+++ filters/BiQuadFilter.cpp
@@ -1,20 +1,20 @@
 /*
-	This file is part of EqualizerAPO, a system-wide equalizer.
-	Copyright (C) 2014  Jonas Thedering
+    This file is part of EqualizerAPO, a system-wide equalizer.
+    Copyright (C) 2014  Jonas Thedering
 
-	This program is free software; you can redistribute it and/or modify
-	it under the terms of the GNU General Public License as published by
-	the Free Software Foundation; either version 2 of the License, or
-	(at your option) any later version.
+    This program is free software; you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation; either version 2 of the License, or
+    (at your option) any later version.
 
-	This program is distributed in the hope that it will be useful,
-	but WITHOUT ANY WARRANTY; without even the implied warranty of
-	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-	GNU General Public License for more details.
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
 
-	You should have received a copy of the GNU General Public License along
-	with this program; if not, write to the Free Software Foundation, Inc.,
-	51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+    You should have received a copy of the GNU General Public License along
+    with this program; if not, write to the Free Software Foundation, Inc.,
+    51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 */
 
 #include "stdafx.h"
@@ -50,7 +50,7 @@ vector<wstring> BiQuadFilter::initialize(float sampleRate, unsigned maxFrameCoun
 		if (!isBandwidthOrS) // Q
 		{
 			double q = bandwidthOrQOrS;
-			double a = pow(10.0, dbGain / 40.0);
+			double a = pow(10, dbGain / 40);
 			s = 1.0 / ((1.0 / (q * q) - 2.0) / (a + 1.0 / a) + 1.0);
 		}
 
@@ -71,17 +71,17 @@ vector<wstring> BiQuadFilter::initialize(float sampleRate, unsigned maxFrameCoun
 }
 
 #pragma AVRT_CODE_BEGIN
-void BiQuadFilter::process(double** output, double** input, unsigned frameCount)
+void BiQuadFilter::process(float** output, float** input, unsigned frameCount)
 {
 	for (unsigned i = 0; i < channelCount; i++)
 	{
 		BiQuad bq = biquads[i];
 
-		double* inputChannel = input[i];
-		double* outputChannel = output[i];
+		float* inputChannel = input[i];
+		float* outputChannel = output[i];
 
 		for (unsigned j = 0; j < frameCount; j++)
-			outputChannel[j] = bq.process(inputChannel[j]);
+			outputChannel[j] = (float)bq.process(inputChannel[j]);
 
 		bq.removeDenormals();
 		biquads[i] = bq;
