+++ filters/PreampFilter.h
@@ -27,7 +27,7 @@ class PreampFilter : public IFilter
 public:
 	PreampFilter(double dbGain);
 	std::vector<std::wstring> initialize(float sampleRate, unsigned maxFrameCount, std::vector<std::wstring> channelNames) override;
-	void process(double** output, double** input, unsigned frameCount) override;
+	void process(float** output, float** input, unsigned frameCount) override;
 
 	double getDbGain() const {return dbGain;}
